requirements.txt
Django==5.2.4
djangorestframework==3.16.0
httpx==0.28.1
pdfminer.six==20250506
python-decouple==3.8
